GitHub-issue#253 : Implemented GeoIP processor functionality. Addressed Code review comments Signed-off-by: venkataraopasyavula <venkataraopasyavula@gmail.com>

+++ data-prepper-plugins/geoip-processor/src/main/java/org/opensearch/dataprepper/plugins/processor/databasedownload/HttpDBDownloadService.java
@@ -90,9 +90,11 @@ public class HttpDBDownloadService implements DBSource {
      * @param tarFolder Tar Folder
      */
     private static void deleteTarFolder(String tarFolder) {
-          final File file = new File(tarFolder);
-          DBSource.deleteDirectory(file);
-          file.delete();
+        final File file = new File(tarFolder);
+        DBSource.deleteDirectory(file);
+        if (file.exists()) {
+            file.delete();
+        }
     }
 
     /**
